refactor `nuxt-picture`
